UTF-8 > UTF-16 parser
It may seem a bad idea to spend a day doing utf8<->utf16 conversion
code, but unfortunately I have to do that because RON(text) is defined
as a UTF-8 based format. The plain text mapper (a.k.a. txt) has to
calculate per-symbol diffs (not per-line like cvs/svn/git/diff/patch).
Hence, we need UTF-16 symbols in the model, or the diff will garble
non-Latin texts.