The span notation, finally!
Some CRDT variants work with chunks of text, just to avoid that
per-char metadata overhead. Indeed, our UUID is 16 bytes. So, in
an RGA/CT text each letter should have 32 bytes of metadata. That's
massive!  Unfortunately, that "chunk" would be yet another entity
in the core model. That would complicate things combinatorially.
Chunks have to be split/merged, addressing becomes much more
complex and so on.  On top of that, we also have the requirement
of Merkle hash checks; mutable chunks throw a wrench into that.
RON is so neat and tidy with its immutable ops.

The span notation solves the problem for us: we still see a text
(formally) as a collection of single-letter operations, but our
serialisation format collapses those char spans into strings.  (A
span is a chain of ops with sequential ids, e.g. @1+A :rdt, @2+A
:1+A, @3+A :2+A...)  As long as ids are sequential, we don't have
to mention them explicitly.  Same applies to same-value
sequences, e.g. spans of rm ops.

@~ 'a span of single-codepoint ops, a.k.a. a "spread"';
@1m138P+aM2GPyfTUY :rga ('this is a span of 25 ops!');

@~ 'a span of 25 remove ops, deletes the string above';
@1m139+aM2GPyfTUY :1m138P000P+aM2GPyfTUY rm (25);