TextFrame: spans and string spreads
...somewhat nicer code and a couple more tests.

The state of the code is now much closer to expressing the concept of
"zero-cost metadata".  Unfortunately, I had to compromise on the performance:
the builder demands atoms to be parsed by default. Booo.

Also, the builder state became a tad more complicated - but this part is
optimized to the necessary minimum.